+++ src/main/java/roomescape/service/ThemeService.java
@@ -3,18 +3,22 @@ package roomescape.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.controller.dto.response.ThemeResponse;
+import roomescape.dao.ReservationDAO;
 import roomescape.dao.ThemeDAO;
 import roomescape.domain.Theme;
 import roomescape.exception.custom.ExistedDuplicateValueException;
 import roomescape.exception.custom.NotExistedValueException;
+import roomescape.exception.custom.PharmaceuticalViolationException;
 import roomescape.service.dto.ThemeCreation;
 
 @Service
 public class ThemeService {
 
+    private final ReservationDAO reservationDAO;
     private final ThemeDAO themeDAO;
 
-    public ThemeService(final ThemeDAO themeDAO) {
+    public ThemeService(final ReservationDAO reservationDAO, final ThemeDAO themeDAO) {
+        this.reservationDAO = reservationDAO;
         this.themeDAO = themeDAO;
     }
 
@@ -37,6 +41,10 @@ public class ThemeService {
     }
 
     public void deleteTheme(final long id) {
+        if (reservationDAO.existsByThemeId(id)) {
+            throw new PharmaceuticalViolationException("사용 중인 테마입니다");
+        }
+
         boolean deleted = themeDAO.deleteById(id);
 
         if (!deleted) {
